// src/app/pages/employees/components/employees-table/employees.interface.ts
export interface EmployeesInterface {
   idemployee?: number;
   idemployees?: number[];
   name?: string;
   phone?: string;
   si_user_idsi_user?: number;
   companyunits_idcompanyunits?: number;
   code?: string;
   photo?: string;
   is_deleted?: boolean;
   created_at?: string;
   created_by?: number;
   modified_at?: string;
}