// src/app/pages/si_users/components/si_users-table/si_users.interface.ts
export interface Si_usersInterface {
   idsi_user?: number;
   nombre?: string;
   apmat?: string;
   appat?: string;
   email?: string;
   password?: string;
   si_rol_idsi_rol?: number;
   companyunits_idcompanyunits?: number;
   status?: string;
   code?: string;
   super?: boolean;
   code_at?: string;
   is_deleted?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}